feat(base-app): restyle quick activate
