+++ apps/base-app/src/app/ui/providers/withi18n/en.json
@@ -20,7 +20,7 @@
     "USER_GUIDE": "User guide"
   },
   "QUICK_ACTIVATE": {
-    "QUICK_ACTIVATE": "Quick activate",
+    "QUICK_ACTIVATE": "Quick activate:",
     "THAT_THIS_DOES_CHANGE_YOUR_PROFILE": "note that this does change your profile",
     "FONT": "Font",
     "FONT_SIZE": "Font size",
