+++ apps/base-app/src/app/ui/providers/withi18n/fr.json
@@ -20,7 +20,7 @@
     "USER_GUIDE": "Guide d'utilisation"
   },
   "QUICK_ACTIVATE": {
-    "QUICK_ACTIVATE": "Activation rapide",
+    "QUICK_ACTIVATE": "Activation rapide : ",
     "THAT_THIS_DOES_CHANGE_YOUR_PROFILE": "notez que cela ne modifie pas votre profil",
     "FONT": "Police",
     "FONT_SIZE": "Taille de police",
