+++ apps/base-app/src/features/toolsQuickActivate/index.ts
@@ -1,0 +1,3 @@
+import ToolsQuickActivate from './ui/ToolsQuickActivate.vue'
+
+export { ToolsQuickActivate }
